chore: move shared helpers into root helper file
